UserVariable holds an Object now instead of a double. * Stringfunctions can use UserVariables as parameter. * The 3 stringfunctions (length(), letter(), join()) can now handle   parameters of the type UserVariable. * Implementation of tests.

+++ catroid/src/org/catrobat/catroid/content/actions/ChangeVariableAction.java
@@ -39,7 +39,7 @@ public class ChangeVariableAction extends Action {
 		if (userVariable == null) {
 			return true;
 		}
-		double originalValue = userVariable.getValue();
+		double originalValue = (Double) userVariable.getValue();
 		double value = changeVariable.interpretDouble(sprite);
 		userVariable.setValue(originalValue + value);
 		return true;
